Comment out launching of cowboy in PlugMintProxy supervisor
Launching the supervisor should probably be handled by a consuming
application rather than by this application.  As such we'll be relying
on others to provide this functionality.

Related code is kept in this repository as it helps understand how the
proxy can be used.  Supporting code will be added as we go along.