Non-relative order of connections
The manipulators had the primary connection as the first element of
the connection tuple and the secondary connection as the second
element of the tuple.  They are now reordered so the frontend_conn is
always first and the backend_conn is always second.  We assume this
will be easier to interpret.