Remove optional implementations from behaviour
Until there's something like state to be initialized through the
behaviour, we know of no nice way to ensure the necessary code is
loaded upon startup.  As such we're forcing the methods to be defined
in the behaviour for now, thereby making the supporting system easier
to run in a predictable way.