Fetch body only once for failed requests
When requests fail, due to connection failures, we should try to
repeat the request with the original body.  This was not the case in
our error handling processes.  Because the request can only be read
from the socket once, this caused issues.

We keep pushing the request upwards but assume this strategy may be
altered in the future if request streaming becomes a desired feature.