Bump plug to v0.14
The latest Bandit version requires this. It's probably also a good idea
to keep up our core components and libraries with the latest versions,
so it's probably good to upgrade this anyway.
We might want to do the same with Mint and CAStore, as Mint has had a
major version bump by now.